+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.0.79](https://github.com/VirtoCommerce/vendor-portal/compare/v1.0.78...v1.0.79) (2023-06-07)
+
+
+### Bug Fixes
+
+* prevent signout on blade changes ([42d0c68](https://github.com/VirtoCommerce/vendor-portal/commit/42d0c68d7a51b2c79e2c29aea3e0ee59f45f9aee))
+
+
+### Features
+
+* improved blade-nav ([cc323cf](https://github.com/VirtoCommerce/vendor-portal/commit/cc323cfc3b2bd366c21480cc303a353500803d78))
+* products bulk delete ([092882f](https://github.com/VirtoCommerce/vendor-portal/commit/092882f31011356b4ed988775e3c97005a46822d))
+
+
+
 ## [1.0.78](https://github.com/VirtoCommerce/vendor-portal/compare/v1.0.77...v1.0.78) (2023-06-02)
 
 
@@ -14,6 +29,7 @@
 * common pages moved to shell ([564ac86](https://github.com/VirtoCommerce/vendor-portal/commit/564ac862bed3e914e0612961652e4f91b557990c))
 * error slot removed from blades ([cafdedf](https://github.com/VirtoCommerce/vendor-portal/commit/cafdedfe55ad17fe07eaea3555014b35b37d46e6))
 * login page props ([3936440](https://github.com/VirtoCommerce/vendor-portal/commit/3936440bd0ddead717f1b66948117a8880dd543d))
+* vm-1131 offers bulk delete ([573d018](https://github.com/VirtoCommerce/vendor-portal/commit/573d01889a9f013e93e1282626c63e3433c22eb1))